deployment/prom/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.7.14 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.2.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
contourpy==1.3.2 ; python_version >= "3.12" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
dateparser==1.2.2 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.116.1 ; python_version >= "3.12" and python_version < "4.0"
fonttools==4.59.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.12" and python_version < "4.0"
httmock==1.4.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
kiwisolver==1.4.8 ; python_version >= "3.12" and python_version < "4.0"
matplotlib==3.10.3 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.3.1 ; python_version >= "3.12" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.3.1 ; python_version >= "3.12" and python_version < "4.0"
pillow==11.3.0 ; python_version >= "3.12" and python_version < "4.0"
prometheus-api-client==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.11.7 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.2.3 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
regex==2024.11.6 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.4 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.47.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.14.1 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.12" and python_version < "4.0"
tzlocal==5.3.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.35.0 ; python_version >= "3.12" and python_version < "4.0"
